// angular/tp-pokemon/src/app/models/pokemonType.ts
export enum PTypes {
    NORMAL = "normal",
    GRASS = "Grass",
    WATER = "Water",
    FIRE = "Fire",
    BUG = "Bug",
    FIGHT = "Fight",
    FLY = "Fly",
    PSYCHIC = "Psychic",
    FAIRY = "Fairy",
    STEEL = "Steel",
    ROCK = "Rock",
    GROUND = "Ground",
    DRAGON = "Dragon",
    ICE = "Ice",
    GHOST = "Ghost",
    DARK = "Dark",
    ELECTRIC = "Electric",
    POISON = "Poison"
}